src/requirements.txt
dash
dash-core-components
dash-html-components
dash-renderer
dash-bootstrap-components==1.0.0rc1
plotly
jsondiff
jsonmerge
flatten_json
importlib-metadata
typing_extensions
GitPython
gunicorn==20.0.4
pandas
genson
typesense
dash_cytoscape
dash-extensions
dash-uploader
packaging
missingpy